Cleanup for 64bit Windows Server build

+++ downloadplugin/download2.cpp
@@ -271,7 +271,7 @@ bool DownloadfileThreaded(std::string url,std::string filename, IPipe *pipe, std
                                                         {
                                                                 out.write(&tmpbuf[4+csize2.size()], tmpbuf.size()-4-csize2.size());
                                                                 out.flush();
-                                                                totalbytes+=tmpbuf.size()-4-csize2.size();
+                                                                totalbytes+=(int)(tmpbuf.size()-4-csize2.size());
                                                         }
                                                 }
                                         }
